opt: add array flatten operator
This change introduces an explicit ArrayFlatten operator. Before, we
would always build an ArrayFlatten as an array_agg, but this doesn't
work for types which array_agg doesn't support (and ArrayFlatten does).
In the interest of replicating all the functionality of the heuristic
planner, ArrayFlattens are now handled like this:

* They are built as an explicit ArrayFlatten operator. If the subquery
  is correlated *and* the type is one which ArrayAgg doesn't support,
  an error is raised.
* If the subquery is correlated, a new norm rule
  `NormalizeArrayFlattenToAgg` is triggered, which converts the
  ArrayFlatten to an array_agg.

Thus, the resulting compatibility looks like this:

                   | Correlated | Uncorrelated
----------------------------------------------
array_agg-able     |    yes     |      yes
not array_agg-able |    no      |      yes

Which is fine, as far as support for features supported by the heuristic
planner goes, because it only supports the "Uncorrelated" column,
anyway.  In order to support the bottom-left quadrant,
frontend/execution need to support array_agg for all types (like
tuples).

There's also another significant change in this commit, which is
required for ArrayFlatten to function properly.

Previously, any subquery could always be optimized with the minimum
physical props, because a single row subquery, or an any, or exists
subquery didn't rely on any kind of ordering. This is no longer true
with ArrayFlatten, since it can require an ordering of its input. So
this commit also makes a handful of changes to the optimization
machinery to support requiring an ordering of a child of a scalar.

Release note: None