opt: reject srfs within aggregates
Trying to backport I realized my change in #38554 was insufficient: it
doesn't work in cases where SRFs occur in a scalar context (the sentry
report used the aggregate in a FROM context, which was already shady in
its own right).

We would miss these because we would set new required properties during
replaceSRF, which resulted in us overwriting the old ones. We fix this
by doing a check of the properties in replaceSRF.

Release note: None